Split hook persistence implementation
