+++ crates/defra-agent/src/hook/persistence/prompt_hook.rs
@@ -1,0 +1,429 @@
+use super::*;
+
+impl<M: CompletionModel> PromptHook<M> for DefraSessionHook {
+    async fn on_completion_call(&self, prompt: &Message, _history: &[Message]) -> HookAction {
+        let result: anyhow::Result<()> = async {
+            let mut state = self.state.lock().await;
+
+            if !state.initialized {
+                let session_id = session::create_session(&self.node, &state.agent_name).await?;
+                state.session_id = Some(session_id);
+                state.initialized = true;
+            }
+
+            state.reset_after_user_message();
+            drop(state);
+
+            self.persist_message(prompt).await?;
+            Ok(())
+        }
+        .await;
+
+        match result {
+            Ok(()) => {
+                self.record_success();
+                HookAction::Continue
+            }
+            Err(e) => self.on_persistence_error("persist user prompt", &e),
+        }
+    }
+
+    async fn on_completion_response(
+        &self,
+        _prompt: &Message,
+        response: &CompletionResponse<M::Response>,
+    ) -> HookAction {
+        let result: anyhow::Result<()> = async {
+            let message = Message::Assistant {
+                id: response.message_id.clone(),
+                content: response.choice.clone(),
+            };
+            let sequence = self.persist_message(&message).await?;
+            self.mark_current_response_materialized(sequence).await?;
+            Ok(())
+        }
+        .await;
+
+        match result {
+            Ok(()) => {
+                self.record_success();
+                HookAction::Continue
+            }
+            Err(e) => self.on_persistence_error("persist assistant response", &e),
+        }
+    }
+
+    async fn on_tool_call(
+        &self,
+        tool_name: &str,
+        tool_call_id: Option<String>,
+        internal_call_id: &str,
+        args: &str,
+    ) -> ToolCallHookAction {
+        if tool_name == SPAWN_SUBAGENT_TOOL_NAME {
+            let result = self
+                .persist_spawn_subagent_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist spawn_subagent tool call", &e),
+            };
+        }
+        if tool_name == WAIT_SUBAGENT_TOOL_NAME {
+            let result = self
+                .persist_wait_subagent_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist wait_subagent tool call", &e),
+            };
+        }
+        if tool_name == LIST_SUBAGENTS_TOOL_NAME {
+            let result = self
+                .persist_list_subagents_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist list_subagents tool call", &e),
+            };
+        }
+        if tool_name == READ_SUBAGENT_TRANSCRIPT_TOOL_NAME {
+            let result = self
+                .persist_read_subagent_transcript_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => {
+                    self.on_tool_persistence_error("persist read_subagent_transcript tool call", &e)
+                }
+            };
+        }
+        if tool_name == STEER_SUBAGENT_TOOL_NAME {
+            let result = self
+                .persist_steer_subagent_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist steer_subagent tool call", &e),
+            };
+        }
+        if tool_name == CANCEL_SUBAGENT_TOOL_NAME {
+            let result = self
+                .persist_cancel_subagent_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist cancel_subagent tool call", &e),
+            };
+        }
+        if tool_name == BACKGROUND_TOOL_NAME {
+            let result = self
+                .persist_background_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist background_tool tool call", &e),
+            };
+        }
+        if tool_name == WAIT_TOOL_NAME {
+            let result = self
+                .persist_wait_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist wait_tool tool call", &e),
+            };
+        }
+        if tool_name == LIST_BACKGROUND_TOOLS_TOOL_NAME {
+            let result = self
+                .persist_list_background_tools_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => {
+                    self.on_tool_persistence_error("persist list_background_tools tool call", &e)
+                }
+            };
+        }
+        if tool_name == READ_TOOL_OUTPUT_TOOL_NAME {
+            let result = self
+                .persist_read_tool_output_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist read_tool_output tool call", &e),
+            };
+        }
+        if tool_name == CANCEL_TOOL_NAME {
+            let result = self
+                .persist_cancel_tool_call(tool_call_id, internal_call_id, args)
+                .instrument(tracing::info_span!(
+                    "tool.call",
+                    tool_name = %tool_name,
+                    tool_call_id = %internal_call_id,
+                ))
+                .await;
+
+            return match result {
+                Ok(action) => {
+                    self.record_success();
+                    action
+                }
+                Err(e) => self.on_tool_persistence_error("persist cancel_tool tool call", &e),
+            };
+        }
+
+        let result: anyhow::Result<()> = async {
+            let (session_id, request_id, deadline_at, seq) =
+                self.ensure_assistant_turn_sequence().await?;
+            self.state.lock().await.register_tool_result_identity(
+                internal_call_id,
+                None,
+                tool_call_id.as_deref(),
+            );
+
+            let mut lc = crate::tool_call_lifecycle::ToolCallLifecycle::new(
+                self.node.clone(),
+                request_id,
+                session_id,
+                internal_call_id.to_string(),
+                seq,
+                tool_name.to_string(),
+                args.to_string(),
+                deadline_at,
+            );
+            lc.start_running().await?;
+
+            self.in_flight_lifecycles
+                .lock()
+                .await
+                .insert(internal_call_id.to_string(), lc);
+
+            Ok(())
+        }
+        .instrument(tracing::info_span!(
+            "tool.call",
+            tool_name = %tool_name,
+            tool_call_id = %internal_call_id,
+        ))
+        .await;
+
+        match result {
+            Ok(()) => {
+                self.record_success();
+                ToolCallHookAction::Continue
+            }
+            Err(e) => self.on_tool_persistence_error("persist tool call", &e),
+        }
+    }
+
+    async fn on_tool_result(
+        &self,
+        tool_name: &str,
+        tool_call_id: Option<String>,
+        internal_call_id: &str,
+        args: &str,
+        result: &str,
+    ) -> HookAction {
+        let persist_result: anyhow::Result<HookAction> = async {
+            if let Some(terminal) = classify_managed_tool_result(result) {
+                let lifecycle = self
+                    .in_flight_lifecycles
+                    .lock()
+                    .await
+                    .remove(internal_call_id);
+
+                if let Some(mut lc) = lifecycle {
+                    match terminal {
+                        ManagedToolTerminal::TimedOut => lc.timeout().await?,
+                        ManagedToolTerminal::Cancelled => lc.cancel_during_run().await?,
+                    }
+                } else {
+                    tracing::debug!(
+                        tool_call_id = %internal_call_id,
+                        lifecycle_state = ?terminal,
+                        "managed terminal tool result arrived after lifecycle was already swept"
+                    );
+                }
+
+                let reason = match terminal {
+                    ManagedToolTerminal::TimedOut => "tool call deadline exceeded",
+                    ManagedToolTerminal::Cancelled => "tool call cancelled",
+                };
+                return Ok(HookAction::Terminate {
+                    reason: reason.to_string(),
+                });
+            }
+
+            let (session_id, should_persist_message, persisted_result_id, persisted_call_id) = {
+                let mut state = self.state.lock().await;
+                let session_id = state
+                    .session_id
+                    .clone()
+                    .ok_or_else(|| anyhow::anyhow!("session hook missing session id"))?;
+                let should_persist_message = state.mark_tool_result_seen_for_persisted_turn(
+                    internal_call_id,
+                    None,
+                    tool_call_id.as_deref(),
+                );
+                let (persisted_result_id, persisted_call_id) =
+                    state.tool_result_message_identity(internal_call_id, tool_call_id.as_deref());
+                (
+                    session_id,
+                    should_persist_message,
+                    persisted_result_id,
+                    persisted_call_id,
+                )
+            };
+
+            let truncator =
+                DefraSpillTruncator::new(self.node.clone(), &self.agent_did, &session_id);
+            let truncated = truncator
+                .truncate(
+                    tool_name,
+                    args,
+                    result,
+                    truncation_mode_for(tool_name),
+                    &self.truncation_limits,
+                    None,
+                )
+                .await?;
+
+            let mut lc = self
+                .in_flight_lifecycles
+                .lock()
+                .await
+                .remove(internal_call_id)
+                .ok_or_else(|| {
+                    anyhow::anyhow!(
+                        "on_tool_result: no in-flight lifecycle for tool_call_id={internal_call_id}"
+                    )
+                })?;
+
+            if let Some(failure_class) = classify_runtime_failure(result) {
+                lc.fail(&truncated.text, failure_class).await?;
+            } else {
+                lc.complete(&truncated.text).await?;
+            }
+
+            if should_persist_message {
+                let tool_result_message = Message::User {
+                    content: OneOrMany::one(UserContent::ToolResult(ToolResult {
+                        id: persisted_result_id,
+                        call_id: persisted_call_id,
+                        content: OneOrMany::one(ToolResultContent::Text(Text {
+                            text: truncated.text.clone(),
+                        })),
+                    })),
+                };
+                self.persist_message(&tool_result_message).await?;
+            }
+
+            Ok(HookAction::Continue)
+        }
+        .instrument(tracing::info_span!(
+            "tool.result",
+            tool_name = %tool_name,
+            tool_call_id = %internal_call_id,
+        ))
+        .await;
+
+        match persist_result {
+            Ok(action) => {
+                self.record_success();
+                action
+            }
+            Err(e) => self.on_persistence_error("persist tool result", &e),
+        }
+    }
+}
